Move files to correct locations
